hf_deployment/requirements.txt
# PlateVision AI - Hugging Face Spaces Requirements
# Optimized for 16GB CPU instance

streamlit==1.28.0
ultralytics==8.0.0
opencv-python-headless==4.8.0.74
pandas==2.0.3
numpy==1.24.3
scipy==1.11.1
filterpy==1.4.5
torch==2.0.1
torchvision==0.15.2
pillow==9.5.0
lapx==0.5.2
matplotlib==3.7.2
